Refactor code
* Revise hasProperty and apply it everywhere
* Add hasIndexedProperty for performance
* Change Object::nextIndex{Forward, Backward} double parameter types into int for performance
* in String.prototype.replace, Use test fast path correctly & use old method because newer method cannot support regexp correctly(old method is faster)

+++ src/runtime/GlobalObjectBuiltinTypedArray.cpp
@@ -677,7 +677,7 @@ static Value builtinTypedArrayLastIndexOf(ExecutionState& state, Value thisValue
                 return Value(k);
             }
         } else {
-            double result;
+            int64_t result;
             Object::nextIndexBackward(state, O, k, -1, false, result);
             k = result;
             continue;
